requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pyswarms>=1.3
pants>=0.6
simanneal>=0.5
dwave-ocean-sdk>=6.0
